# docs/swagger.yaml
definitions:
  httputil.HTTPError:
    properties:
      code:
        example: 400
        type: integer
      message:
        example: status bad request
        type: string
    type: object
  models.Aluno:
    properties:
      cpf:
        type: string
      nome:
        type: string
    required:
    - cpf
    - nome
    type: object
info:
  contact: {}
paths:
  /:
    get:
      consumes:
      - application/json
      description: objetivo é retornar uma lista de aluno atraves
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Aluno'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httputil.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.HTTPError'
      summary: Lista de Alunos
      tags:
      - Aluno
  /alunos/{id}:
    delete:
      consumes:
      - application/json
      description: objetivo é deletar um aluno atraves do seu id
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Aluno'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httputil.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.HTTPError'
      summary: Deletar aluno
      tags:
      - Aluno
    get:
      consumes:
      - application/json
      description: objetivo é retornar uma aluno atraves do seu id
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Aluno'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httputil.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.HTTPError'
      summary: Buscar um Aluno Por Id
      tags:
      - Aluno
    patch:
      consumes:
      - application/json
      description: Objetivo é atualizar um Aluno existente
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: integer
      - description: Add Aluno
        in: body
        name: Aluno
        required: true
        schema:
          $ref: '#/definitions/models.Aluno'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Aluno'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httputil.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.HTTPError'
      summary: Atualizar um Aluno
      tags:
      - Aluno
  /alunos/novo:
    post:
      consumes:
      - application/json
      description: Objetivo é cadastra um novo Aluno
      parameters:
      - description: Add Aluno
        in: body
        name: Aluno
        required: true
        schema:
          $ref: '#/definitions/models.Aluno'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Aluno'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httputil.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.HTTPError'
      summary: Novo Aluno
      tags:
      - Aluno
swagger: "2.0"